debug nodemailer and cloudinary error

+++ backend/controllers/movies.js
@@ -1,4 +1,5 @@
-const Movie = require('../models/movies')
+const Movie = require('../models/movies');
+const { uploadImage, uploadVideo, deleteImageFromCloudinary, deleteVideoFromCloudinary } = require('../utils/uploadImgVid');
 
 async function movieAddHandler(req, res) {
   try {
@@ -11,21 +12,27 @@ async function movieAddHandler(req, res) {
     }
 
 
-    const image = req.files['image']?.[0]?.path;
-    const video = req.files['video']?.[0]?.path;
+    const imagePath = req.files['image']?.[0]?.path;
+    const videoPath = req.files['video']?.[0]?.path;
 
-    if (!image || !video) {
+    if (!imagePath || !videoPath) {
       return res.status(400).json({ message: 'Both image and video are required' });
     }
 
+    const imageResult = await uploadImage(imagePath);
+    const videoResult = await uploadVideo(videoPath);
+
+    if (!imageResult || !videoResult) {
+      return res.status(500).json({ message: 'Failed to upload files to Cloudinary' });
+    }
 
     const newMovie = new Movie({
       name,
       director,
       releasedDate,
       duration,
-      image,
-      video,
+      image: imageResult.secure_url,
+      video: videoResult.secure_url,
     });
 
     await newMovie.save();
@@ -89,8 +96,25 @@ async function getMovieById(req,res) {
 }
 
 async function deleteMovieHandler(req, res) {
-  await Movie.findByIdAndDelete(req.params.id)
-  return res.status(204).json({ "message": "Deleted successfully" })
+  try {
+    const movie = await Movie.findById(req.params.id);
+    if (!movie) {
+      return res.status(404).json({ message: 'Movie not found' });
+    }
+
+    const imagePublicId = movie.image.split('/').slice(-2).join('/').split('.')[0];
+    const videoPublicId = movie.video.split('/').slice(-2).join('/').split('.')[0];
+
+    await deleteImageFromCloudinary(imagePublicId);
+    await deleteVideoFromCloudinary(videoPublicId);
+
+    await Movie.findByIdAndDelete(req.params.id);
+    
+    return res.status(200).json({ message: 'Movie deleted successfully' });
+  } catch (error) {
+    console.error(error);
+    return res.status(500).json({ message: 'Internal Server Error' });
+  }
 }
 
 async function countMovieHandler(req,res) {
